requirements.txt
dnspython==2.8.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
pydantic>=2.7.0
mcp>=1.0.0
